+++ packages/lib/location.ts
@@ -1,0 +1,10 @@
+export enum LocationType {
+  InPerson = "inPerson",
+  Phone = "phone",
+  GoogleMeet = "integrations:google:meet",
+  Zoom = "integrations:zoom",
+  Daily = "integrations:daily",
+  Jitsi = "integrations:jitsi",
+  Huddle01 = "integrations:huddle01",
+  Tandem = "integrations:tandem",
+}
